chore(ECO-2982): Address PR comments
