+++ src/typescript/frontend/src/components/pages/emojicoin/components/trade-emojicoin/LiquidityButton.tsx
@@ -22,7 +22,6 @@ export const LiquidityButton = (props: GridProps) => {
         <StyledContentHeader>
           <Flex width="100%" justifyContent="center">
             <Link
-              prefetch={false}
               href={{
                 pathname: ROUTES.pools,
                 query: { pool: props.data.emojis.map((e) => e.emoji).join("") },
