+++ src/typescript/frontend/src/components/pages/home/components/emoji-table/index.tsx
@@ -172,7 +172,7 @@ const EmojiTable = (props: EmojiTableProps) => {
               </>
             ) : (
               <div className="py-10">
-                <Link prefetch={false} href={`${ROUTES.launch}?emojis=${emojis.join("")}`}>
+                <Link href={`${ROUTES.launch}?emojis=${emojis.join("")}`}>
                   <Text textScale="pixelHeading3" color="econiaBlue" className="uppercase">
                     Click here to launch {<Emoji emojis={emojis.join("")} />}
                   </Text>
