+++ src/typescript/frontend/src/components/pages/home/components/main-card/MainCard.tsx
@@ -83,7 +83,6 @@ const MainCard = (props: MainCardProps) => {
     <div className="flex flex-col w-full my-[20px] md:my-[70px] max-w-full">
       <div className="flex flex-col md:flex-row w-full max-w-full items-center justify-center">
         <Link
-          prefetch={false}
           className="flex relative items-center ml-[-8%]"
           href={
             featured
