+++ src/typescript/frontend/src/components/price-feed/inner.tsx
@@ -33,7 +33,6 @@ export const PriceDelta = ({ delta, className = "" }: { delta: number; className
 const Item = ({ emoji, delta }: { emoji: string; delta: number }) => {
   return (
     <Link
-      prefetch={false}
       href={`/market/${emoji}`}
       className={`font-pixelar whitespace-nowrap border-[1px] border-solid ${delta >= 0 ? "border-green" : "border-pink"} rounded-full px-3 py-[2px] select-none mr-[22px]`}
       draggable={false}
